istio being optional
