+++ controllers/kuadrant_controller.go
@@ -23,23 +23,16 @@ import (
 	"github.com/go-logr/logr"
 	authorinov1beta1 "github.com/kuadrant/authorino-operator/api/v1beta1"
 	limitadorv1alpha1 "github.com/kuadrant/limitador-operator/api/v1alpha1"
-	iopv1alpha1 "istio.io/istio/operator/pkg/apis/istio/v1alpha1"
 	appsv1 "k8s.io/api/apps/v1"
-	corev1 "k8s.io/api/core/v1"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
-	apimeta "k8s.io/apimachinery/pkg/api/meta"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/utils/env"
-	istiov1alpha1 "maistra.io/istio-operator/api/v1alpha1"
 	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/controller/controllerutil"
 
-	maistrav1 "github.com/kuadrant/kuadrant-operator/api/external/maistra/v1"
-	maistrav2 "github.com/kuadrant/kuadrant-operator/api/external/maistra/v2"
 	kuadrantv1beta1 "github.com/kuadrant/kuadrant-operator/api/v1beta1"
 	"github.com/kuadrant/kuadrant-operator/pkg/common"
-	"github.com/kuadrant/kuadrant-operator/pkg/istio"
+	kuadrantistioutils "github.com/kuadrant/kuadrant-operator/pkg/istio"
 	"github.com/kuadrant/kuadrant-operator/pkg/kuadranttools"
 	"github.com/kuadrant/kuadrant-operator/pkg/library/reconcilers"
 	"github.com/kuadrant/kuadrant-operator/pkg/log"
@@ -47,8 +40,6 @@ import (
 
 const (
 	kuadrantFinalizer = "kuadrant.io/finalizer"
-	// (Sail) The istio CR must be named default to process GW API resources
-	istioCRName = "default"
 )
 
 // KuadrantReconciler reconciles a Kuadrant object
@@ -103,7 +94,7 @@ func (r *KuadrantReconciler) Reconcile(eventCtx context.Context, req ctrl.Reques
 	if kObj.GetDeletionTimestamp() != nil && controllerutil.ContainsFinalizer(kObj, kuadrantFinalizer) {
 		logger.V(1).Info("Handling removal of kuadrant object")
 
-		if err := r.unregisterExternalAuthorizer(ctx, kObj); err != nil {
+		if err := kuadrantistioutils.UnregisterExternalAuthorizer(ctx, r.Client(), kObj.Namespace); err != nil {
 			return ctrl.Result{}, err
 		}
 
@@ -149,217 +140,8 @@ func (r *KuadrantReconciler) Reconcile(eventCtx context.Context, req ctrl.Reques
 	return ctrl.Result{}, nil
 }
 
-func (r *KuadrantReconciler) unregisterExternalAuthorizer(ctx context.Context, kObj *kuadrantv1beta1.Kuadrant) error {
-	logger, _ := logr.FromContext(ctx)
-
-	isIstioInstalled, err := r.unregisterExternalAuthorizerIstio(ctx, kObj)
-
-	if err == nil && !isIstioInstalled {
-		err = r.unregisterExternalAuthorizerOSSM(ctx, kObj)
-	}
-
-	if err != nil {
-		logger.Error(err, "failed fo get service mesh control plane")
-	}
-
-	return err
-}
-
-func (r *KuadrantReconciler) unregisterExternalAuthorizerIstio(ctx context.Context, kObj *kuadrantv1beta1.Kuadrant) (bool, error) {
-	logger, _ := logr.FromContext(ctx)
-	configsToUpdate, err := r.getIstioConfigObjects(ctx, logger)
-	isIstioInstalled := configsToUpdate != nil
-
-	if !isIstioInstalled || err != nil {
-		return isIstioInstalled, err
-	}
-
-	kuadrantAuthorizer := common.NewKuadrantAuthorizer(kObj.Namespace)
-
-	for _, config := range configsToUpdate {
-		hasKuadrantAuthorizer, err := common.HasKuadrantAuthorizer(config, *kuadrantAuthorizer)
-		if err != nil {
-			return true, err
-		}
-		if hasKuadrantAuthorizer {
-			if err = common.UnregisterKuadrantAuthorizer(config, kuadrantAuthorizer); err != nil {
-				return true, err
-			}
-
-			logger.Info("remove external authorizer from istio meshconfig")
-			if err = r.UpdateResource(ctx, config.GetConfigObject()); err != nil {
-				return true, err
-			}
-		}
-	}
-	return true, nil
-}
-
-func (r *KuadrantReconciler) unregisterExternalAuthorizerOSSM(ctx context.Context, kObj *kuadrantv1beta1.Kuadrant) error {
-	logger, _ := logr.FromContext(ctx)
-
-	smcp := &maistrav2.ServiceMeshControlPlane{}
-
-	smcpKey := client.ObjectKey{Name: controlPlaneProviderName(), Namespace: controlPlaneProviderNamespace()}
-	if err := r.Client().Get(ctx, smcpKey, smcp); err != nil {
-		logger.V(1).Info("failed to get servicemeshcontrolplane object", "key", smcp, "err", err)
-		return err
-	}
-
-	smcpWrapper := istio.NewOSSMControlPlaneWrapper(smcp)
-	kuadrantAuthorizer := common.NewKuadrantAuthorizer(kObj.Namespace)
-
-	hasKuadrantAuthorizer, err := common.HasKuadrantAuthorizer(smcpWrapper, *kuadrantAuthorizer)
-	if err != nil {
-		return err
-	}
-	if hasKuadrantAuthorizer {
-		err = common.UnregisterKuadrantAuthorizer(smcpWrapper, kuadrantAuthorizer)
-		if err != nil {
-			return err
-		}
-		logger.Info("removing external authorizer from  OSSM meshconfig")
-		if err := r.UpdateResource(ctx, smcpWrapper.GetConfigObject()); err != nil {
-			return err
-		}
-	}
-
-	return nil
-}
-
-func (r *KuadrantReconciler) registerExternalAuthorizer(ctx context.Context, kObj *kuadrantv1beta1.Kuadrant) error {
-	logger, _ := logr.FromContext(ctx)
-
-	isIstioInstalled, err := r.registerExternalAuthorizerIstio(ctx, kObj)
-
-	if err == nil && !isIstioInstalled {
-		err = r.registerExternalAuthorizerOSSM(ctx, kObj)
-	}
-
-	if err != nil {
-		logger.Error(err, "failed fo get service mesh control plane")
-	}
-
-	return err
-}
-
-func (r *KuadrantReconciler) registerExternalAuthorizerIstio(ctx context.Context, kObj *kuadrantv1beta1.Kuadrant) (bool, error) {
-	logger, _ := logr.FromContext(ctx)
-	configsToUpdate, err := r.getIstioConfigObjects(ctx, logger)
-	isIstioInstalled := configsToUpdate != nil
-
-	if !isIstioInstalled || err != nil {
-		return isIstioInstalled, err
-	}
-
-	kuadrantAuthorizer := common.NewKuadrantAuthorizer(kObj.Namespace)
-	for _, config := range configsToUpdate {
-		hasKuadrantAuthorizer, err := common.HasKuadrantAuthorizer(config, *kuadrantAuthorizer)
-		if err != nil {
-			return true, err
-		}
-		if !hasKuadrantAuthorizer {
-			err = common.RegisterKuadrantAuthorizer(config, kuadrantAuthorizer)
-			if err != nil {
-				return true, err
-			}
-			logger.Info("adding external authorizer to istio meshconfig")
-			if err = r.UpdateResource(ctx, config.GetConfigObject()); err != nil {
-				return true, err
-			}
-		}
-	}
-
-	return true, nil
-}
-
-func (r *KuadrantReconciler) registerExternalAuthorizerOSSM(ctx context.Context, kObj *kuadrantv1beta1.Kuadrant) error {
-	logger, _ := logr.FromContext(ctx)
-
-	if err := r.registerServiceMeshMember(ctx, kObj); err != nil {
-		return err
-	}
-
-	smcp := &maistrav2.ServiceMeshControlPlane{}
-
-	smcpKey := client.ObjectKey{Name: controlPlaneProviderName(), Namespace: controlPlaneProviderNamespace()}
-	if err := r.GetResource(ctx, smcpKey, smcp); err != nil {
-		logger.V(1).Info("failed to get servicemeshcontrolplane object", "key", smcp, "err", err)
-		return err
-	}
-	smcpWrapper := istio.NewOSSMControlPlaneWrapper(smcp)
-	kuadrantAuthorizer := common.NewKuadrantAuthorizer(kObj.Namespace)
-
-	hasKuadrantAuthorizer, err := common.HasKuadrantAuthorizer(smcpWrapper, *kuadrantAuthorizer)
-	if err != nil {
-		return err
-	}
-	if !hasKuadrantAuthorizer {
-		err = common.RegisterKuadrantAuthorizer(smcpWrapper, kuadrantAuthorizer)
-		if err != nil {
-			return err
-		}
-		logger.Info("adding external authorizer to OSSM meshconfig")
-		if err := r.UpdateResource(ctx, smcpWrapper.GetConfigObject()); err != nil {
-			return err
-		}
-	}
-
-	return nil
-}
-
-func (r *KuadrantReconciler) getIstioConfigObjects(ctx context.Context, logger logr.Logger) ([]common.ConfigWrapper, error) {
-	var configsToUpdate []common.ConfigWrapper
-
-	iop := &iopv1alpha1.IstioOperator{}
-	istKey := client.ObjectKey{Name: controlPlaneProviderName(), Namespace: controlPlaneProviderNamespace()}
-	err := r.GetResource(ctx, istKey, iop)
-	if err == nil || apierrors.IsNotFound(err) {
-		configsToUpdate = append(configsToUpdate, istio.NewOperatorWrapper(iop))
-	} else if !apimeta.IsNoMatchError(err) {
-		logger.V(1).Info("failed to get istiooperator object", "key", istKey, "err", err)
-		return nil, err
-	} else {
-		// Error is NoMatchError so check for Istio CR instead
-		ist := &istiov1alpha1.Istio{}
-		istKey := client.ObjectKey{Name: istioCRName}
-		if err := r.GetResource(ctx, istKey, ist); err != nil {
-			logger.V(1).Info("failed to get istio object", "key", istKey, "err", err)
-			if apimeta.IsNoMatchError(err) {
-				// return nil and nil if there's no istiooperator or istio CR
-				return nil, nil
-			} else if !apierrors.IsNotFound(err) {
-				// return nil and err if there's an error other than not found (no istio CR)
-				return nil, err
-			}
-		}
-		configsToUpdate = append(configsToUpdate, istio.NewSailWrapper(ist))
-	}
-
-	istioConfigMap := &corev1.ConfigMap{}
-	if err := r.GetResource(ctx, client.ObjectKey{Name: controlPlaneConfigMapName(), Namespace: controlPlaneProviderNamespace()}, istioConfigMap); err != nil {
-		if !apierrors.IsNotFound(err) {
-			logger.V(1).Info("failed to get istio configMap", "key", istKey, "err", err)
-			return configsToUpdate, err
-		}
-	} else {
-		configsToUpdate = append(configsToUpdate, istio.NewConfigMapWrapper(istioConfigMap))
-	}
-	return configsToUpdate, nil
-}
-
-func (r *KuadrantReconciler) registerServiceMeshMember(ctx context.Context, kObj *kuadrantv1beta1.Kuadrant) error {
-	member := buildServiceMeshMember(kObj)
-	err := r.SetOwnerReference(kObj, member)
-	if err != nil {
-		return err
-	}
-
-	return r.ReconcileResource(ctx, &maistrav1.ServiceMeshMember{}, member, reconcilers.CreateOnlyMutator)
-}
-
 func (r *KuadrantReconciler) reconcileSpec(ctx context.Context, kObj *kuadrantv1beta1.Kuadrant) error {
-	if err := r.registerExternalAuthorizer(ctx, kObj); err != nil {
+	if err := kuadrantistioutils.RegisterExternalAuthorizer(ctx, r.Client(), kObj, r.Scheme()); err != nil {
 		return err
 	}
 
@@ -368,37 +150,6 @@ func (r *KuadrantReconciler) reconcileSpec(ctx context.Context, kObj *kuadrantv1
 	}
 
 	return r.reconcileAuthorino(ctx, kObj)
-}
-
-func controlPlaneProviderName() string {
-	return env.GetString("ISTIOOPERATOR_NAME", "istiocontrolplane")
-}
-
-func controlPlaneConfigMapName() string {
-	return env.GetString("ISTIOCONFIGMAP_NAME", "istio")
-}
-
-func controlPlaneProviderNamespace() string {
-	return env.GetString("ISTIOOPERATOR_NAMESPACE", "istio-system")
-}
-
-func buildServiceMeshMember(kObj *kuadrantv1beta1.Kuadrant) *maistrav1.ServiceMeshMember {
-	return &maistrav1.ServiceMeshMember{
-		TypeMeta: metav1.TypeMeta{
-			Kind:       "ServiceMeshMember",
-			APIVersion: maistrav1.SchemeGroupVersion.String(),
-		},
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "default",
-			Namespace: kObj.Namespace,
-		},
-		Spec: maistrav1.ServiceMeshMemberSpec{
-			ControlPlaneRef: maistrav1.ServiceMeshControlPlaneRef{
-				Name:      controlPlaneProviderName(),
-				Namespace: controlPlaneProviderNamespace(),
-			},
-		},
-	}
 }
 
 func (r *KuadrantReconciler) reconcileLimitador(ctx context.Context, kObj *kuadrantv1beta1.Kuadrant) error {
