+++ make/envoy-gateway.mk
@@ -26,7 +26,7 @@ egctl: $(EGCTL) ## Download egctl locally if necessary.
 
 EG_VERSION ?= v0.0.0-latest
 .PHONY: envoy-gateway-install
-envoy-gateway-install: kustomize
+envoy-gateway-install: kustomize $(HELM)
 	$(HELM) install eg oci://docker.io/envoyproxy/gateway-helm --version $(EG_VERSION) -n envoy-gateway-system --create-namespace
 	kubectl wait --timeout=5m -n envoy-gateway-system deployment/envoy-gateway --for=condition=Available
