+++ pkg/istio/mesh_config_test.go
@@ -15,6 +15,7 @@ import (
 	corev1 "k8s.io/api/core/v1"
 	istiov1alpha1 "maistra.io/istio-operator/api/v1alpha1"
 	"maistra.io/istio-operator/pkg/helm"
+	"sigs.k8s.io/controller-runtime/pkg/client"
 )
 
 func getStubbedMeshConfig() *istiomeshv1alpha1.MeshConfig {
@@ -264,4 +265,88 @@ func TestSailWrapper_SetMeshConfig(t *testing.T) {
 
 	assert.Equal(t, meshConfig.ExtensionProviders[0].Name, stubbedMeshConfig.ExtensionProviders[0].Name)
 	assert.Equal(t, meshConfig.ExtensionProviders[0].GetEnvoyExtAuthzGrpc().GetPort(), uint32(50051))
+}
+
+func getStubbedMeshConfig() *istiomeshv1alpha1.MeshConfig {
+	providers := make([]*istiomeshv1alpha1.MeshConfig_ExtensionProvider, 0)
+	provider := &istiomeshv1alpha1.MeshConfig_ExtensionProvider{
+		Name: "custom-authorizer",
+		Provider: &istiomeshv1alpha1.MeshConfig_ExtensionProvider_EnvoyExtAuthzGrpc{
+			EnvoyExtAuthzGrpc: &istiomeshv1alpha1.MeshConfig_ExtensionProvider_EnvoyExternalAuthorizationGrpcProvider{
+				Port:    50051,
+				Service: "custom-authorizer.default.svc.cluster.local",
+			},
+		},
+	}
+	providers = append(providers, provider)
+	return &istiomeshv1alpha1.MeshConfig{
+		ExtensionProviders: providers,
+	}
+}
+
+type stubbedConfigWrapper struct {
+	istioMeshConfig *istiomeshv1alpha1.MeshConfig
+}
+
+func (c *stubbedConfigWrapper) SetMeshConfig(config *istiomeshv1alpha1.MeshConfig) error {
+	c.istioMeshConfig = config
+	return nil
+}
+
+func (c *stubbedConfigWrapper) GetMeshConfig() (*istiomeshv1alpha1.MeshConfig, error) {
+	return c.istioMeshConfig, nil
+}
+
+func (c *stubbedConfigWrapper) GetConfigObject() client.Object {
+	return nil
+}
+
+func TestKuadrantAuthorizer_GetExtensionProvider(t *testing.T) {
+	authorizer := NewKuadrantAuthorizer("default")
+	provider := authorizer.GetExtensionProvider()
+
+	assert.Equal(t, provider.Name, ExtAuthorizerName)
+	assert.Equal(t, provider.GetEnvoyExtAuthzGrpc().Service, "authorino-authorino-authorization.default.svc.cluster.local")
+}
+
+func TestHasKuadrantAuthorizer(t *testing.T) {
+	authorizer := NewKuadrantAuthorizer("default")
+	configWrapper := &stubbedConfigWrapper{getStubbedMeshConfig()}
+
+	hasAuthorizer, err := HasKuadrantAuthorizer(configWrapper, *authorizer)
+
+	assert.NilError(t, err)
+	assert.Equal(t, hasAuthorizer, false)
+
+	configWrapper.istioMeshConfig.ExtensionProviders = append(configWrapper.istioMeshConfig.ExtensionProviders, authorizer.GetExtensionProvider())
+	hasAuthorizer, err = HasKuadrantAuthorizer(configWrapper, *authorizer)
+	assert.NilError(t, err)
+	assert.Equal(t, hasAuthorizer, true)
+}
+
+func TestRegisterKuadrantAuthorizer(t *testing.T) {
+	authorizer := NewKuadrantAuthorizer("default")
+	configWrapper := &stubbedConfigWrapper{getStubbedMeshConfig()}
+
+	err := RegisterKuadrantAuthorizer(configWrapper, authorizer)
+	assert.NilError(t, err)
+
+	meshConfig, _ := configWrapper.GetMeshConfig()
+	assert.Equal(t, meshConfig.ExtensionProviders[1].Name, "kuadrant-authorization")
+}
+
+func TestUnregisterKuadrantAuthorizer(t *testing.T) {
+	authorizer := NewKuadrantAuthorizer("default")
+	configWrapper := &stubbedConfigWrapper{getStubbedMeshConfig()}
+
+	err := RegisterKuadrantAuthorizer(configWrapper, authorizer)
+	assert.NilError(t, err)
+	assert.Equal(t, len(configWrapper.istioMeshConfig.ExtensionProviders), 2)
+
+	err = UnregisterKuadrantAuthorizer(configWrapper, authorizer)
+	assert.NilError(t, err)
+	assert.Equal(t, len(configWrapper.istioMeshConfig.ExtensionProviders), 1)
+
+	meshConfig, _ := configWrapper.GetMeshConfig()
+	assert.Equal(t, meshConfig.GetExtensionProviders()[0].Name, "custom-authorizer")
 }
