+++ pkg/istio/utils.go
@@ -5,6 +5,10 @@ import (
 
 	"github.com/go-logr/logr"
 	istiocommon "istio.io/api/type/v1beta1"
+	istioclientgoextensionv1alpha1 "istio.io/client-go/pkg/apis/extensions/v1alpha1"
+	istioclientnetworkingv1alpha3 "istio.io/client-go/pkg/apis/networking/v1alpha3"
+	"k8s.io/apimachinery/pkg/api/meta"
+	"k8s.io/apimachinery/pkg/runtime/schema"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	gatewayapiv1 "sigs.k8s.io/gateway-api/apis/v1"
 
@@ -21,4 +25,38 @@ func WorkloadSelectorFromGateway(ctx context.Context, k8sClient client.Client, g
 	return &istiocommon.WorkloadSelector{
 		MatchLabels: gatewayWorkloadSelector,
 	}
+}
+
+func IsIstioWASMPluginInstalled(restMapper meta.RESTMapper) (bool, error) {
+	_, err := restMapper.RESTMapping(
+		schema.GroupKind{Group: istioclientgoextensionv1alpha1.GroupName, Kind: "WasmPlugin"},
+		istioclientgoextensionv1alpha1.SchemeGroupVersion.Version,
+	)
+
+	if err == nil {
+		return true, nil
+	}
+
+	if meta.IsNoMatchError(err) {
+		return false, nil
+	}
+
+	return false, err
+}
+
+func IsIstioEnvoyFilterInstalled(restMapper meta.RESTMapper) (bool, error) {
+	_, err := restMapper.RESTMapping(
+		schema.GroupKind{Group: istioclientnetworkingv1alpha3.GroupName, Kind: "EnvoyFilter"},
+		istioclientnetworkingv1alpha3.SchemeGroupVersion.Version,
+	)
+
+	if err == nil {
+		return true, nil
+	}
+
+	if meta.IsNoMatchError(err) {
+		return false, nil
+	}
+
+	return false, err
 }
